Reorganize CustomDevice class hierarchy
